UNet/requirements.txt
matplotlib==3.6.2
numpy==1.23.5
Pillow==9.3.0
tqdm==4.64.1
torch~=2.1.0+cu118
torchvision~=0.16.0+cu118
scikit-learn~=1.3.2
wandb~=0.13.5